+++ frontend/src/components/player/Card.jsx
@@ -1,29 +1,8 @@
-import { useEffect, useState } from "react";
+import '@/styles/components/neonCard.css'
 
-function Card({ card, setSelectedCards, selectedCards }) {
-  const [isFlipped, setIsFlipped] = useState(false);
-  const handleClick = () => {
-    setSelectedCards([...selectedCards, card]);
-  };
-  useEffect(() => {
-    if (
-      selectedCards[0] === card ||
-      selectedCards[1] === card ||
-      card.ismatch
-    ) {
-      setIsFlipped(true);
-    } else {
-      setIsFlipped(false);
-    }
-  }, [selectedCards]);
-  return (
-    <div className={isFlipped ? "card open" : "card"} onClick={handleClick}>
-      <div className="front">
-        <h1 className="animalName">{card.animal}</h1>
-      </div>
-      <div className="back"></div>
-    </div>
-  );
+const NeonCard = () =>{
+    return(
+        <button className="flip cardStyle"><p className="tit">?</p></button>
+    )
 }
-
-export default Card;
+export default NeonCard;
